+++ NEWS.md
@@ -1,6 +1,8 @@
 # CHANGES IN pagedown VERSION 0.8
 
+## BUG FIXES
 
+- In `chrome_print()`, fixed some connection problems to headless Chrome: in some situations, the R session tries to connect to headless Chrome before a target is created. Now, `chrome_print()` controls the target creation by connecting to the `Browser` endpoint (thanks, @gershomtripp, #158).  
 
 # CHANGES IN pagedown VERSION 0.7